+++ src/DiskIO/DiskDaemon/diskd.cc
@@ -1,5 +1,5 @@
 /*
- * $Id: diskd.cc,v 1.7 2007/04/25 16:41:56 wessels Exp $
+ * $Id: diskd.cc,v 1.8 2007/08/16 23:32:28 hno Exp $
  *
  * DEBUG: section --    External DISKD process implementation.
  * AUTHOR: Harvest Derived
@@ -150,11 +150,11 @@ do_read(diomsg * r, int len, char *buf)
 
     if (r->offset > -1 && r->offset != fs->offset) {
         DEBUG(2)
-        fprintf(stderr, "seeking to %d\n", r->offset);
+        fprintf(stderr, "seeking to %"PRId64"\n", (int64_t)r->offset);
 
         if (lseek(fs->fd, r->offset, SEEK_SET) < 0) {
             DEBUG(1) {
-                fprintf(stderr, "%d FD %d, offset %d: ", (int) mypid, fs->fd, r->offset);
+                fprintf(stderr, "%d FD %d, offset %"PRId64": ", (int) mypid, fs->fd, (int64_t)r->offset);
                 perror("lseek");
             }
         }
@@ -162,8 +162,8 @@ do_read(diomsg * r, int len, char *buf)
 
     x = read(fs->fd, buf, readlen);
     DEBUG(2)
-    fprintf(stderr, "%d READ %d,%d,%d ret %d\n", (int) mypid,
-            fs->fd, readlen, r->offset, x);
+    fprintf(stderr, "%d READ %d,%d,%"PRId64" ret %d\n", (int) mypid,
+            fs->fd, readlen, (int64_t)r->offset, x);
 
     if (x < 0) {
         DEBUG(1) {
@@ -199,15 +199,15 @@ do_write(diomsg * r, int len, const char *buf)
     if (r->offset > -1 && r->offset != fs->offset) {
         if (lseek(fs->fd, r->offset, SEEK_SET) < 0) {
             DEBUG(1) {
-                fprintf(stderr, "%d FD %d, offset %d: ", (int) mypid, fs->fd, r->offset);
+                fprintf(stderr, "%d FD %d, offset %"PRId64": ", (int) mypid, fs->fd, (int64_t)r->offset);
                 perror("lseek");
             }
         }
     }
 
     DEBUG(2)
-    fprintf(stderr, "%d WRITE %d,%d,%d\n", (int) mypid,
-            fs->fd, wrtlen, r->offset);
+    fprintf(stderr, "%d WRITE %d,%d,%"PRId64"\n", (int) mypid,
+            fs->fd, wrtlen, (int64_t)r->offset);
     x = write(fs->fd, buf, wrtlen);
 
     if (x < 0) {
