+++ src/DiskIO/DiskThreads/aiops_win32.cc
@@ -1,5 +1,5 @@
 /*
- * $Id: aiops_win32.cc,v 1.4 2007/04/28 22:26:47 hno Exp $
+ * $Id: aiops_win32.cc,v 1.5 2007/08/16 23:32:28 hno Exp $
  *
  * DEBUG: section 43    Windows AIOPS
  * AUTHOR: Stewart Forster <slf@connect.com.au>
@@ -72,7 +72,7 @@ typedef struct squidaio_request_t
     int fd;
     char *bufferp;
     char *tmpbufp;
-    int buflen;
+    size_t buflen;
     off_t offset;
     int whence;
     int ret;
@@ -788,7 +788,7 @@ squidaio_do_open(squidaio_request_t * requestp)
 
 
 int
-squidaio_read(int fd, char *bufp, int bufs, off_t offset, int whence, squidaio_result_t * resultp)
+squidaio_read(int fd, char *bufp, size_t bufs, off_t offset, int whence, squidaio_result_t * resultp)
 {
     squidaio_request_t *requestp;
 
@@ -834,7 +834,7 @@ squidaio_do_read(squidaio_request_t * requestp)
 
 
 int
-squidaio_write(int fd, char *bufp, int bufs, off_t offset, int whence, squidaio_result_t * resultp)
+squidaio_write(int fd, char *bufp, size_t bufs, off_t offset, int whence, squidaio_result_t * resultp)
 {
     squidaio_request_t *requestp;
